requirements.txt
Pillow==9.5.0
torch==2.0.0+cu117
torchvision==0.12.0+cu113
numpy==1.21.1
